+++ src/main/kotlin/io/openfuture/chain/domain/crypto/key/KeyDto.kt
@@ -1,0 +1,6 @@
+package io.openfuture.chain.domain.crypto.key
+
+data class KeyDto(
+    val publicKey: String,
+    val privateKey: String? = null
+)