+++ src/test/kotlin/io/openfuture/chain/config/Mockito.kt
@@ -1,0 +1,5 @@
+package io.openfuture.chain.config
+
+import org.mockito.Mockito
+
+fun <T> any(clazz: Class<T>): T = Mockito.any<T>(clazz)